[jit] define equality for IValue
Previously we were reluctant to define equality for IValues, as it's not
totally straightforward. But the vacuum this created basically forced
people to define their own equality comparisons for their own purposes.
We have at least 3 in PyTorch itself, and 2 others outside that I know
of.

These implementations are generally wrong, so we should just bite the
bullet and define equality canonically.

ghstack-source-id: 3da869ba7c95988956385e4a2df922952d94db8a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34986